Fix “transaction already closed” error when running tasks
