+++ apps/webapp/app/routes/api.v1.runs.$runId.tasks.ts
@@ -2,21 +2,16 @@ import type { ActionFunctionArgs } from "@remix-run/server-runtime";
 import { json } from "@remix-run/server-runtime";
 import {
   API_VERSIONS,
-  RunTaskBodyOutput,
   RunTaskBodyOutputSchema,
   RunTaskResponseWithCachedTasksBody,
   ServerTask,
 } from "@trigger.dev/core";
-import { TaskStatus } from "@trigger.dev/database";
 import { z } from "zod";
-import { $transaction, PrismaClient, prisma } from "~/db.server";
-import { env } from "~/env.server";
-import { prepareTasksForCaching, taskWithAttemptsToServerTask } from "~/models/task.server";
+import { PrismaClient, prisma } from "~/db.server";
+import { prepareTasksForCaching } from "~/models/task.server";
 import { authenticateApiRequest } from "~/services/apiAuth.server";
 import { logger } from "~/services/logger.server";
-import { generateSecret } from "~/services/sources/utils.server";
-import { ulid } from "~/services/ulid.server";
-import { taskOperationWorker, workerQueue } from "~/services/worker.server";
+import { RunTaskService } from "~/services/tasks/runTask.server";
 
 const ParamsSchema = z.object({
   runId: z.string(),
@@ -159,186 +154,5 @@ class ChangeRequestLazyLoadedCachedTasks {
       task,
       cachedTasks: preparedTasks,
     };
-  }
-}
-
-export class RunTaskService {
-  #prismaClient: PrismaClient;
-
-  constructor(prismaClient: PrismaClient = prisma) {
-    this.#prismaClient = prismaClient;
-  }
-
-  public async call(
-    runId: string,
-    idempotencyKey: string,
-    taskBody: RunTaskBodyOutput
-  ): Promise<ServerTask | undefined> {
-    const task = await $transaction(this.#prismaClient, async (tx) => {
-      const existingTask = await tx.task.findUnique({
-        where: {
-          runId_idempotencyKey: {
-            runId,
-            idempotencyKey,
-          },
-        },
-        include: {
-          attempts: true,
-          run: true,
-        },
-      });
-
-      const delayUntilInFuture = taskBody.delayUntil && taskBody.delayUntil.getTime() > Date.now();
-      const callbackEnabled = taskBody.callback?.enabled;
-
-      if (existingTask) {
-        if (existingTask.status === "CANCELED") {
-          const existingTaskStatus =
-            delayUntilInFuture || callbackEnabled || taskBody.trigger
-              ? "WAITING"
-              : taskBody.noop
-              ? "COMPLETED"
-              : "RUNNING";
-
-          const resumedExistingTask = await tx.task.update({
-            where: {
-              id: existingTask.id,
-            },
-            data: {
-              status: existingTaskStatus,
-              startedAt: new Date(),
-              completedAt: existingTaskStatus === "COMPLETED" ? new Date() : undefined,
-            },
-            include: {
-              run: true,
-              attempts: true,
-            },
-          });
-
-          return resumedExistingTask;
-        }
-
-        return existingTask;
-      }
-
-      const run = await tx.jobRun.findUnique({
-        where: {
-          id: runId,
-        },
-        select: {
-          status: true,
-        },
-      });
-
-      if (!run) throw new Error("Run not found");
-
-      // If task.delayUntil is set and is in the future, we'll set the task's status to "WAITING", else set it to RUNNING
-      let status: TaskStatus;
-
-      if (run.status === "CANCELED") {
-        status = "CANCELED";
-      } else {
-        status =
-          delayUntilInFuture || callbackEnabled || taskBody.trigger
-            ? "WAITING"
-            : taskBody.noop
-            ? "COMPLETED"
-            : "RUNNING";
-      }
-
-      const taskId = ulid();
-      const callbackUrl = callbackEnabled
-        ? `${env.APP_ORIGIN}/api/v1/tasks/${taskId}/callback/${generateSecret(12)}`
-        : undefined;
-
-      const task = await tx.task.create({
-        data: {
-          id: taskId,
-          idempotencyKey,
-          displayKey: taskBody.displayKey,
-          runConnection: taskBody.connectionKey
-            ? {
-                connect: {
-                  runId_key: {
-                    runId,
-                    key: taskBody.connectionKey,
-                  },
-                },
-              }
-            : undefined,
-          icon: taskBody.icon,
-          run: {
-            connect: {
-              id: runId,
-            },
-          },
-          parent: taskBody.parentId ? { connect: { id: taskBody.parentId } } : undefined,
-          name: taskBody.name ?? "Task",
-          description: taskBody.description,
-          status,
-          startedAt: new Date(),
-          completedAt: status === "COMPLETED" || status === "CANCELED" ? new Date() : undefined,
-          noop: taskBody.noop,
-          delayUntil: taskBody.delayUntil,
-          params: taskBody.params ?? undefined,
-          properties: this.#filterProperties(taskBody.properties) ?? undefined,
-          redact: taskBody.redact ?? undefined,
-          operation: taskBody.operation,
-          callbackUrl,
-          style: taskBody.style ?? { style: "normal" },
-          childExecutionMode: taskBody.parallel ? "PARALLEL" : "SEQUENTIAL",
-          attempts: {
-            create: {
-              number: 1,
-              status: "PENDING",
-            },
-          },
-        },
-        include: {
-          run: true,
-          attempts: true,
-        },
-      });
-
-      if (task.status === "RUNNING" && typeof taskBody.operation === "string") {
-        // We need to schedule the operation
-        await taskOperationWorker.enqueue(
-          "performTaskOperation",
-          {
-            id: task.id,
-          },
-          { tx, runAt: task.delayUntil ?? undefined, jobKey: `operation:${task.id}` }
-        );
-      } else if (task.status === "WAITING" && callbackUrl && taskBody.callback) {
-        if (taskBody.callback.timeoutInSeconds > 0) {
-          // We need to schedule the callback timeout
-          await workerQueue.enqueue(
-            "processCallbackTimeout",
-            {
-              id: task.id,
-            },
-            {
-              tx,
-              runAt: new Date(Date.now() + taskBody.callback.timeoutInSeconds * 1000),
-              jobKey: `process-callback:${task.id}`,
-            }
-          );
-        }
-      }
-
-      return task;
-    });
-
-    return task ? taskWithAttemptsToServerTask(task) : undefined;
-  }
-
-  #filterProperties(properties: RunTaskBodyOutput["properties"]): RunTaskBodyOutput["properties"] {
-    if (!properties) return;
-
-    return properties.filter((property) => {
-      if (!property) return false;
-
-      return typeof property.label === "string" && typeof property.text === "string";
-    });
   }
 }
